Backend/requirements.txt
Flask==2.3.3
flask-bcrypt==1.0.1
pymongo==4.5.0
python-dotenv==1.0.0
pyrebase4==4.7.1
requests>=2.19.1,<2.30